requirements.txt
pytesseract
PyPDF2
fastapi
uvicorn
PyMuPDF
pillow
prisma
google-generativeai
poppler-utils
